excel/pyproject.toml
[project]
name = "excel"
version = "0.1.0"
description = ""
authors = [
    {name = "luigi ",email = "150992024@eniac.edu.br"}
]
requires-python =">=3.11"
dependencies = [
    "scrapy (>=2.14.2,<3.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "setuptools (==70.0.0)",
    "playwright-stealth (==1.0.6)",
    "scrapy-playwright (>=0.0.46,<0.0.47)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "openai (>=2.30.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
